Move grayscale sample extraction
